extraction of repoInit so successful git think's it's a rename
